+++ pyscriptjs/environment.yml
@@ -9,3 +9,4 @@ dependencies:
   - black
   - isort
   - codespell
+  - pre-commit
