+++ pyscriptjs/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "svelte-app",
-  "version": "1.0.0",
+  "name": "pyscript",
+  "version": "0.0.1",
   "scripts": {
     "build": "NODE_ENV=production rollup -c",
     "dev": "rollup -c -w",
@@ -15,36 +15,36 @@
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
-    "@rollup/plugin-typescript": "^8.1.0",
+    "@rollup/plugin-typescript": "^8.3.2",
     "@tsconfig/svelte": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^5.20.0",
     "@typescript-eslint/parser": "^5.20.0",
-    "autoprefixer": "^10.2.3",
+    "autoprefixer": "^10.4.7",
     "eslint": "^8.14.0",
     "eslint-plugin-svelte3": "^3.4.1",
-    "postcss": "^8.2.4",
+    "postcss": "^8.4.13",
     "prettier": "^2.6.2",
     "prettier-plugin-svelte": "^2.7.0",
-    "rollup": "^2.3.4",
+    "rollup": "^2.71.1",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-livereload": "^2.0.0",
     "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-svelte": "^7.0.0",
     "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.0.0",
+    "svelte": "^3.48.0",
     "svelte-check": "^1.0.0",
-    "svelte-preprocess": "^4.6.3",
+    "svelte-preprocess": "^4.10.6",
     "tailwindcss": "^2.0.2",
-    "tslib": "^2.0.0",
-    "typescript": "^4.1.3"
+    "tslib": "^2.4.0",
+    "typescript": "^4.6.4"
   },
   "dependencies": {
     "@codemirror/basic-setup": "^0.19.1",
-    "@codemirror/lang-python": "^0.19.4",
+    "@codemirror/lang-python": "^0.19.5",
     "@codemirror/state": "^0.19.9",
     "@codemirror/theme-one-dark": "^0.19.1",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "codemirror": "^5.65.2",
+    "codemirror": "^5.65.3",
     "js-yaml": "^4.1.0",
     "sirv-cli": "^1.0.0",
     "svelte-fa": "^2.4.0",
